Provider of data for registries at Rapporteket

Generic to registries, provide reporting data obtained from sql databases Underlying this function is rapbase::RapporteketDbConnection

LoadRegData(registryName, query, dbType = "mysql")

Arguments

registryName
String Name of the registry as defined in dbConfig.yml

query
String SQL query to obtain the data

dbType
String Type of db to query, currently "mysql" (default) and "mssql"

Value

RegData dataframe Registry data