Get configuration for package, if any

Try to obtain yaml-formatted configuration placed either as given by the environment variable R_RAP_CONFIG_PATH or as provided by the package itself. If none can be found the function exits with an error

getConfig(fileName = "dbConfig.yml", packageName = "rapbase")

Arguments

fileName
String providing configuration file base name

packageName
String providing the package name

Value

A list of (yaml) configuration

Examples

getConfig()

#> $rapbase
#> $rapbase$host
#> [1] "db"
#> 
#> $rapbase$name
#> [1] "rapbase"
#> 
#> $rapbase$user
#> [1] "root"
#> 
#> $rapbase$pass
#> [1] "root"
#> 
#> $rapbase$disp
#> [1] "ForTestingOnly"
#> 
#> 
#> $smerte
#> $smerte$host
#> [1] "db"
#> 
#> $smerte$name
#> [1] "SmerteReportDataStaging"
#> 
#> $smerte$user
#> [1] "root"
#> 
#> $smerte$pass
#> [1] "root"
#> 
#> $smerte$disp
#> [1] "Smerteregisteret"
#> 
#> 
#> $nir
#> $nir$host
#> [1] "db"
#> 
#> $nir$name
#> [1] "IntensivRegisterReportDataStaging"
#> 
#> $nir$user
#> [1] "root"
#> 
#> $nir$pass
#> [1] "root"
#> 
#> $nir$disp
#> [1] "Intensivregisteret"
#> 
#> 
#> $norgast
#> $norgast$host
#> [1] "db"
#> 
#> $norgast$name
#> [1] "NoRGastReportDataStaging"
#> 
#> $norgast$user
#> [1] "root"
#> 
#> $norgast$pass
#> [1] "root"
#> 
#> $norgast$disp
#> [1] "NoRGast"
#> 
#> 
#> $noricStaging102966
#> $noricStaging102966$host
#> [1] "db"
#> 
#> $noricStaging102966$name
#> [1] "NORICReportDataStagingBergen"
#> 
#> $noricStaging102966$user
#> [1] "root"
#> 
#> $noricStaging102966$pass
#> [1] "root"
#> 
#> $noricStaging102966$disp
#> [1] "Bergen"
#> 
#> 
#> $Nakke
#> $Nakke$host
#> [1] "db"
#> 
#> $Nakke$name
#> [1] "DegenNakkeReportDataStaging"
#> 
#> $Nakke$user
#> [1] "root"
#> 
#> $Nakke$pass
#> [1] "root"
#> 
#> $Nakke$disp
#> [1] "Nakke"
#> 
#>