Get user full name from config or session object

This is a helper function for userInfo. When used without a shiny session object calls to this function is made without any arguments. If redefining contexts is needed, please use userInfo instead.

getUserFullName(shinySession = NULL)

Arguments

shinySession
Shiny session object (list, NULL by default). Must be provided when the source of user attributes is either the shiny app url or an external authentication provider. By default this will apply to the 'TEST', 'QA' and 'PRODUCTION' contexts in which case the shiny session object must be provided.

Value

String full name

See also

Examples

# NOT RUN {
getUserFullName()
getUserFullName(shinySessionObject)
# }