Get user name from config or session object

This is a helper function for userInfo. When used without a shiny session object calls to this function is made without any arguments. If redefining contexts is needed, please use userInfo instead.

getUserName(shinySession = NULL)

Arguments

shinySession

Shiny session object (list, NULL by default). Must be provided when the source of user attributes is either the shiny app url or an external authentication provider. By default this will apply to the 'TEST', 'QA' and 'PRODUCTION' contexts in which case the shiny session object must be provided.

Value

String user name

See also

Examples

# \donttest{
# Requires a valid shiny session object
try(getUserName())
#> System has no defined instance. Configuration as provided by
#>             'rapbaseConfig.yml' will be used as source for user data.
#> [1] "ttester"
try(getUserName(shinySessionObject))
#> System has no defined instance. Configuration as provided by
#>             'rapbaseConfig.yml' will be used as source for user data.
#> [1] "ttester"
# }