Management of shiny application and report logging at Rapporteket

A tool to aid archiving and cleanup of logs containing data on use of shiny application and reports at Rapporteket. This function is ment to be called from a scheduler (e.g. chron) directly or as part of a process chain initiated by a scheduler

raplogManager(overSize = 1024 * 1000)

Arguments

overSize
Integer size in bytes from where larger files will be listed as candidates for archiving. Not used by this function per se but rather passed on to logsOverSize(). Normally, its value does not need to be set but may be convenient in rare cases (such as full scale tests). Default value set to 1 Mb (1024 * 1000)

Value

Silently exits after successful file operations

See also

Examples

# \donttest{
# Depend on the R_RAP_CONFIG_PATH being defined and that relevant
# configuration is present
try(raplogManager())

#> Error in raplogManager() : 
#>   No path to configuration data provided. Cannot continue!

# }