Read automated report metadata

Read automated report metadata

readAutoReportData(fileName = "autoReport.yml", packageName = "rapbase")

Arguments

fileName
String defining name of the yaml configuration file. Default 'autoReport.yml'

packageName
String defining the package in which the above configuration file resides. A configuration file within an R-package is only used in case the environmental variable 'R_RAP_CONFIG_PATH' is not defined (empty)

Value

a list of yaml data

Examples

readAutoReportData()
#> Auto report data were upgraded: auto reports with no type defined now set to 'subscription'.
#> Auto report data were upgraded: auto reports with no owner name defined now set to an empty string.
#> Auto report data were upgraded: function params list un-nested. Please check that autor reports for registries are still working as expected.
#> $testAutoReportFirst
#> $testAutoReportFirst$synopsis
#> [1] "Test, the first one"
#> 
#> $testAutoReportFirst$package
#> [1] "rapbase"
#> 
#> $testAutoReportFirst$type
#> [1] "dispatchment"
#> 
#> $testAutoReportFirst$fun
#> [1] ".testAutoReport"
#> 
#> $testAutoReportFirst$params
#> $testAutoReportFirst$params$aNum
#> [1] 1
#> 
#> $testAutoReportFirst$params$aChar
#> [1] "a"
#> 
#> $testAutoReportFirst$params$anExp
#> [1] "Sys.Date()"
#> 
#> 
#> $testAutoReportFirst$owner
#> [1] "ttester"
#> 
#> $testAutoReportFirst$email
#> [1] "Some One <someone@nowhere.com>" "Jesus <jesus@sky.com>"         
#> 
#> $testAutoReportFirst$organization
#> [1] "999999"
#> 
#> $testAutoReportFirst$startDate
#> [1] "1900-01-01"
#> 
#> $testAutoReportFirst$terminateDate
#> [1] "9999-12-31"
#> 
#> $testAutoReportFirst$interval
#> NULL
#> 
#> $testAutoReportFirst$intervalName
#> NULL
#> 
#> $testAutoReportFirst$runDayOfYear
#> [1] 30 60
#> 
#> $testAutoReportFirst$ownerName
#> [1] ""
#> 
#> 
#> $testAutoReportSecond
#> $testAutoReportSecond$synopsis
#> [1] "Test, the second one"
#> 
#> $testAutoReportSecond$package
#> [1] "rapbase"
#> 
#> $testAutoReportSecond$type
#> [1] "subscription"
#> 
#> $testAutoReportSecond$fun
#> [1] ".testAutoReport"
#> 
#> $testAutoReportSecond$params
#> $testAutoReportSecond$params$aNum
#> [1] 2
#> 
#> $testAutoReportSecond$params$aChar
#> [1] "3.14"
#> 
#> $testAutoReportSecond$params$anExp
#> [1] "sessionInfo()"
#> 
#> 
#> $testAutoReportSecond$owner
#> [1] "ttester"
#> 
#> $testAutoReportSecond$email
#> [1] "<someone@nowhere.com>"
#> 
#> $testAutoReportSecond$organization
#> [1] "999999"
#> 
#> $testAutoReportSecond$startDate
#> [1] "1900-01-01"
#> 
#> $testAutoReportSecond$terminateDate
#> [1] "9999-12-31"
#> 
#> $testAutoReportSecond$interval
#> NULL
#> 
#> $testAutoReportSecond$intervalName
#> NULL
#> 
#> $testAutoReportSecond$runDayOfYear
#> [1] 30 60 90
#> 
#> $testAutoReportSecond$ownerName
#> [1] ""
#> 
#> 
#> $testAutoReportThird
#> $testAutoReportThird$synopsis
#> [1] "Test, the second one"
#> 
#> $testAutoReportThird$type
#> [1] "bulletin"
#> 
#> $testAutoReportThird$package
#> [1] "rapbase"
#> 
#> $testAutoReportThird$fun
#> [1] ".testAutoReport"
#> 
#> $testAutoReportThird$params
#> $testAutoReportThird$params$aNum
#> [1] 2
#> 
#> $testAutoReportThird$params$aChar
#> [1] "3.14"
#> 
#> $testAutoReportThird$params$anExp
#> [1] "sessionInfo()"
#> 
#> $testAutoReportThird$params$bulletin
#> [1] TRUE
#> 
#> 
#> $testAutoReportThird$owner
#> [1] "ttester"
#> 
#> $testAutoReportThird$ownerName
#> [1] "Tore Tester"
#> 
#> $testAutoReportThird$email
#> [1] "<someone@nowhere.com>" "jesus@sky.com"        
#> 
#> $testAutoReportThird$organization
#> [1] "999999"
#> 
#> $testAutoReportThird$startDate
#> [1] "1900-01-01"
#> 
#> $testAutoReportThird$terminateDate
#> [1] "9999-12-31"
#> 
#> $testAutoReportThird$interval
#> NULL
#> 
#> $testAutoReportThird$intervalName
#> NULL
#> 
#> $testAutoReportThird$runDayOfYear
#> [1] 30 60 90
#> 
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$synopsis
#> [1] "Test of auto report"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$package
#> [1] "rapbase"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$fun
#> [1] ".testAutoReport"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$params
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$params$aNum
#> [1] "1"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$params$aChar
#> [1] "someString"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$params$anExp
#> [1] "Sys.Date()"
#> 
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$owner
#> [1] "ttester"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$email
#> [1] "tester@skde.no"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$organization
#> [1] "999999"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$startDate
#> [1] "1900-01-01"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$terminateDate
#> [1] "9999-12-31"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$interval
#> NULL
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$intervalName
#> NULL
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$runDayOfYear
#> [1] 170
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$type
#> [1] "subscription"
#> 
#> $`5a0e7e28fd4c641be535a2b292a6b7b2`$ownerName
#> [1] ""
#> 
#> 
#> $cd467ac14dd848b8798a384a3e51512c
#> $cd467ac14dd848b8798a384a3e51512c$synopsis
#> [1] "Test of expired auto report"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$package
#> [1] "rapbase"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$fun
#> [1] ".testAutoReport"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$params
#> $cd467ac14dd848b8798a384a3e51512c$params$aNum
#> [1] "1"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$params$aChar
#> [1] "someString"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$params$anExp
#> [1] "Sys.Date()"
#> 
#> 
#> $cd467ac14dd848b8798a384a3e51512c$owner
#> [1] "ttester"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$email
#> [1] "tester@skde.no"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$organization
#> [1] "999999"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$startDate
#> [1] "1900-01-01"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$terminateDate
#> [1] "0000-01-01"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$interval
#> NULL
#> 
#> $cd467ac14dd848b8798a384a3e51512c$intervalName
#> NULL
#> 
#> $cd467ac14dd848b8798a384a3e51512c$runDayOfYear
#> [1] 171
#> 
#> $cd467ac14dd848b8798a384a3e51512c$type
#> [1] "subscription"
#> 
#> $cd467ac14dd848b8798a384a3e51512c$ownerName
#> [1] ""
#> 
#>