Find next run date for automated reports

Find the next date that an automated report is supposed to be run. Likely, this function will only be relevant for shiny apps when this date is to be printed

findNextRunDate(
  runDayOfYear,
  baseDayNum = as.POSIXlt(Sys.Date())$yday + 1,
  returnFormat = "%A %e. %B %Y"
)

Arguments

runDayOfYear
Numeric vector providing year-day numbers

baseDayNum
Numeric defining base year-day. Default is today

returnFormat
String providing return format as described in strptime in the current locale. Defaults to "%A %d. %B %Y" that will provide something like 'Mandag 20. januar 2019' in a Norwegian locale

Value

String date for printing

Examples

# Will return Jan 30 in the current year and locale with default formatting
findNextRunDate(c(10, 20, 30), 20)

#> [1] "Saturday 30. January 2021"