Read automated report metadata

Read automated report metadata

readAutoReportData(fileName = "autoReport.yml",
  packageName = "rapbase")

Arguments

fileName
String defining name of the yaml configuration file. Default 'autoReport.yml'

packageName
String defining the package in which the above configuration file resides. A configuration file within an R-package is only used in case the environmental variable 'R_RAP_CONFIG_PATH' is not defined (empty)

Value

a list of yaml data

Examples

readAutoReportData()

#> $testAutoReportFirst
#> $testAutoReportFirst$synopsis
#> [1] "Test, the first one"
#> 
#> $testAutoReportFirst$package
#> [1] "rapbase"
#> 
#> $testAutoReportFirst$fun
#> [1] ".testAutoReport"
#> 
#> $testAutoReportFirst$params
#> $testAutoReportFirst$params[[1]]
#> $testAutoReportFirst$params[[1]]$aNum
#> [1] 1
#> 
#> 
#> $testAutoReportFirst$params[[2]]
#> $testAutoReportFirst$params[[2]]$aChar
#> [1] "a"
#> 
#> 
#> $testAutoReportFirst$params[[3]]
#> $testAutoReportFirst$params[[3]]$anExp
#> [1] "Sys.Date()"
#> 
#> 
#> 
#> $testAutoReportFirst$owner
#> [1] "tester"
#> 
#> $testAutoReportFirst$email
#> [1] "Some One <someone@nowhere.com>" "Jesus <jesus@sky.com>"         
#> 
#> $testAutoReportFirst$organization
#> [1] 0
#> 
#> $testAutoReportFirst$runDayOfYear
#> [1] 30 60
#> 
#> 
#> $testAutoReportSecond
#> $testAutoReportSecond$synopsis
#> [1] "Test, the second one"
#> 
#> $testAutoReportSecond$package
#> [1] "rapbase"
#> 
#> $testAutoReportSecond$fun
#> [1] ".testAutoReport"
#> 
#> $testAutoReportSecond$params
#> $testAutoReportSecond$params[[1]]
#> $testAutoReportSecond$params[[1]]$aNum
#> [1] 2
#> 
#> 
#> $testAutoReportSecond$params[[2]]
#> $testAutoReportSecond$params[[2]]$aChar
#> [1] "3.14"
#> 
#> 
#> $testAutoReportSecond$params[[3]]
#> $testAutoReportSecond$params[[3]]$anExp
#> [1] "sessionInfo()"
#> 
#> 
#> 
#> $testAutoReportSecond$owner
#> [1] "tester"
#> 
#> $testAutoReportSecond$email
#> [1] "<someone@nowhere.com>" "jesus@sky.com"        
#> 
#> $testAutoReportSecond$organization
#> [1] 0
#> 
#> $testAutoReportSecond$runDayOfYear
#> [1] 30 60 90
#> 
#>